misc. onboarding improvements
(cherry picked from commit c5fabccb07f0c223f42df2ffc7d943197091e28f)
