+++ apps/daimo-mobile/src/view/screen/onboarding/OnboardingScreen.tsx
@@ -1,21 +1,13 @@
 import { DaimoLinkInvite, DaimoLinkNoteV2, assertNotNull } from "@daimo/common";
 import { DaimoChain } from "@daimo/contract";
 import Octicons from "@expo/vector-icons/Octicons";
-import * as Device from "expo-device";
 import { addEventListener, getInitialURL } from "expo-linking";
-import * as Notifications from "expo-notifications";
-import { ReactNode, useEffect, useRef, useState } from "react";
-import {
-  ActivityIndicator,
-  NativeScrollEvent,
-  NativeSyntheticEvent,
-  Platform,
-  ScrollView,
-  StyleSheet,
-  View,
-} from "react-native";
+import { useEffect, useRef, useState } from "react";
+import { ActivityIndicator, Platform, StyleSheet, View } from "react-native";
 
+import { AllowNotificationsPage } from "./AllowNotificationsPage";
 import { CreateAccountPage } from "./CreateAccountPage";
+import { IntroPages } from "./IntroPages";
 import { InvitePage } from "./InvitePage";
 import { OnboardingHeader } from "./OnboardingHeader";
 import { UseExistingPage } from "./UseExistingPage";
@@ -24,11 +16,8 @@ import { useCreateAccount } from "../../../action/useCreateAccount";
 import { getInvitePasteLink } from "../../../logic/invite";
 import { requestEnclaveSignature } from "../../../logic/key";
 import { NamedError } from "../../../logic/log";
-import { getPushNotificationManager } from "../../../logic/notify";
 import { defaultEnclaveKeyName } from "../../../model/account";
 import { ButtonBig } from "../../shared/Button";
-import { InfoLink } from "../../shared/InfoLink";
-import { IntroTextParagraph } from "../../shared/IntroTextParagraph";
 import Spacer from "../../shared/Spacer";
 import { color, ss } from "../../shared/style";
 import {
@@ -36,13 +25,11 @@ import {
   TextBody,
   TextCenter,
   TextError,
-  TextH1,
   TextLight,
 } from "../../shared/text";
 
 type OnboardPage =
   | "intro"
-  | "flow-selection"
   | "create-invite"
   | "create-try-enclave"
   | "existing-try-enclave"
@@ -66,7 +53,7 @@ export default function OnboardingScreen({
     setPage(newPage);
   };
   const goToPrev = () => pageStack.length > 0 && setPage(pageStack.pop()!);
-  const [daimoChain, setDaimoChain] = useState<DaimoChain>("base");
+  const [daimoChain, setDaimoChain] = useState<DaimoChain>("baseGoerli");
 
   const next = getNext(page, goTo, setDaimoChain, onOnboardingComplete);
   const prev = pageStack.length === 0 ? undefined : goToPrev;
@@ -108,7 +95,7 @@ export default function OnboardingScreen({
 
   const reset = () => {
     pageStack.length = 0;
-    goTo("flow-selection");
+    goTo("intro");
     setName("");
     setInviteLink(undefined);
     createReset && createReset();
@@ -117,7 +104,6 @@ export default function OnboardingScreen({
   return (
     <View style={styles.onboardingScreen}>
       {page === "intro" && <IntroPages onNext={next} />}
-      {page === "flow-selection" && <FlowSelectionPage onNext={next} />}
       {page === "create-invite" && (
         <InvitePage
           onNext={next}
@@ -178,8 +164,6 @@ function getNext(
   const fnGoTo = (p: OnboardPage) => () => goToPage(p);
   switch (page) {
     case "intro":
-      return fnGoTo("flow-selection");
-    case "flow-selection":
       return (input) => {
         const { choice } = assertNotNull(input);
         // Android goes through an extra onboarding step
@@ -214,154 +198,6 @@ function getNext(
     default:
       throw new Error(`unreachable ${page}`);
   }
-}
-
-const tokenSymbol = "USDC";
-const introPages = [
-  <IntroPage title="Welcome to Daimo">
-    <IntroTextParagraph>
-      Daimo is a global payments app that runs on Ethereum. Send and receive
-      USDC on Base mainnet.
-    </IntroTextParagraph>
-  </IntroPage>,
-  <IntroPage title={tokenSymbol}>
-    <IntroTextParagraph>
-      You can send and receive money using the {tokenSymbol} stablecoin. 1{" "}
-      {tokenSymbol} is $1.
-    </IntroTextParagraph>
-    <View style={ss.container.marginHNeg16}>
-      <InfoLink
-        url="https://www.circle.com/en/usdc"
-        title="Learn how it works here"
-      />
-    </View>
-  </IntroPage>,
-  <IntroPage title="Yours alone">
-    <IntroTextParagraph>
-      Daimo stores money via cryptographic secrets. There's no bank.
-    </IntroTextParagraph>
-  </IntroPage>,
-  <IntroPage title="On Ethereum">
-    <IntroTextParagraph>
-      Daimo runs on Base, an Ethereum rollup. This lets you send money securely,
-      anywhere in the world.
-    </IntroTextParagraph>
-  </IntroPage>,
-];
-
-function IntroPages({ onNext }: { onNext: () => void }) {
-  const [pageIndex, setPageIndex] = useState(0);
-  const updatePageBubble = (event: NativeSyntheticEvent<NativeScrollEvent>) => {
-    const { contentOffset, layoutMeasurement } = event.nativeEvent;
-    const page = Math.round(contentOffset.x / layoutMeasurement.width);
-    setPageIndex(page);
-  };
-
-  return (
-    <View style={styles.introPages}>
-      <PageBubble count={4} index={pageIndex} />
-      <ScrollView
-        horizontal
-        pagingEnabled
-        showsHorizontalScrollIndicator={false}
-        style={styles.introPageScroll}
-        onScroll={updatePageBubble}
-        scrollEventThrottle={32}
-        contentContainerStyle={{ width: `${introPages.length * 100}%` }}
-      >
-        {introPages.map((page, i) => (
-          <View style={{ width: `${100 / introPages.length}%` }} key={i}>
-            {page}
-          </View>
-        ))}
-      </ScrollView>
-      <Spacer h={32} />
-      <View style={styles.introButtonsCenter}>
-        <View style={styles.introButtonsWrap}>
-          <ButtonBig type="primary" title="Enter" onPress={onNext} />
-        </View>
-      </View>
-    </View>
-  );
-}
-
-function IntroPage({
-  title,
-  children,
-}: {
-  title: string;
-  children: ReactNode;
-}) {
-  return (
-    <View style={styles.introPage}>
-      <TextCenter>
-        <TextH1>{title}</TextH1>
-      </TextCenter>
-      <Spacer h={32} />
-      {children}
-    </View>
-  );
-}
-
-function PageBubble({ count, index }: { count: number; index: number }) {
-  const bubbles = [];
-  for (let i = 0; i < count; i++) {
-    bubbles.push(
-      <View
-        key={i}
-        style={{
-          width: 8,
-          height: 8,
-          borderRadius: 4,
-          backgroundColor: i === index ? color.midnight : color.grayLight,
-          margin: 4,
-        }}
-      />
-    );
-  }
-  return (
-    <View style={{ flexDirection: "row", justifyContent: "center" }}>
-      {bubbles}
-    </View>
-  );
-}
-
-function FlowSelectionPage({
-  onNext,
-}: {
-  onNext: ({ choice }: { choice: "create" | "existing" }) => void;
-}) {
-  return (
-    <View style={styles.paddedPage}>
-      <Spacer h={32} />
-      <TextCenter>
-        <TextH1>Welcome</TextH1>
-      </TextCenter>
-      <Spacer h={64} />
-      <TextCenter>
-        <IntroTextParagraph>
-          You can create a new account, or add this device to a Daimo account
-          you already have.
-        </IntroTextParagraph>
-      </TextCenter>
-      <Spacer h={32} />
-      <ButtonBig
-        type="primary"
-        title="Create Account"
-        onPress={() => {
-          onNext({ choice: "create" });
-        }}
-      />
-      <Spacer h={16} />
-      <ButtonBig
-        type="subtle"
-        title="Already have an account?"
-        onPress={() => {
-          onNext({ choice: "existing" });
-        }}
-      />
-    </View>
-  );
 }
 
 function SetupKeyPage({
@@ -414,13 +250,13 @@ function SetupKeyPage({
             color={color.midnight}
           />
         </View>
-        <Spacer h={32} />
+        <Spacer h={24} />
         <View style={ss.container.padH16}>
           <TextCenter>
             {!askToSetPin && (
               <TextBody>
-                Generate your Daimo account. Your account is stored on your
-                device, secured by cryptography.
+                Generate your Daimo device key. This key is generated and stored
+                on your device, and secures access to your Daimo account.
               </TextBody>
             )}
             {askToSetPin && (
@@ -431,7 +267,7 @@ function SetupKeyPage({
             )}
           </TextCenter>
         </View>
-        <Spacer h={32} />
+        <Spacer h={118} />
         {(loading || createStatus === "loading") && (
           <ActivityIndicator size="large" />
         )}
@@ -451,46 +287,6 @@ function SetupKeyPage({
           </>
         )}
       </View>
-    </View>
-  );
-}
-
-function AllowNotificationsPage({ onNext }: { onNext: () => void }) {
-  const requestPermission = async () => {
-    if (!Device.isDevice) {
-      window.alert("Push notifications unsupported in simulator.");
-      return;
-    }
-
-    const status = await Notifications.requestPermissionsAsync();
-    console.log(`[ONBOARDING] notifications request ${status.status}`);
-    if (!status.granted) return;
-
-    getPushNotificationManager().maybeSavePushTokenForAccount();
-
-    onNext();
-  };
-
-  return (
-    <View style={styles.paddedPage}>
-      <TextCenter>
-        <Octicons name="bell" size={32} />
-      </TextCenter>
-      <Spacer h={32} />
-      <TextCenter>
-        <IntroTextParagraph>
-          You'll be notified only for account activity. For example, when you
-          receive money.
-        </IntroTextParagraph>
-      </TextCenter>
-      <Spacer h={32} />
-      <ButtonBig
-        type="primary"
-        title="Allow Notifications"
-        onPress={requestPermission}
-      />
-      <Spacer h={16} />
-      <ButtonBig type="subtle" title="Skip" onPress={onNext} />
     </View>
   );
 }
@@ -533,30 +329,6 @@ function CreateAccountSpinnerPage({
 const styles = StyleSheet.create({
   onboardingScreen: {
     flex: 1,
-  },
-  introPages: {
-    flex: 1,
-    backgroundColor: color.white,
-    alignItems: "stretch",
-    justifyContent: "center",
-  },
-  introPageScroll: {
-    flexGrow: 0,
-  },
-  introPage: {
-    padding: 32,
-    maxWidth: 480,
-    alignSelf: "center",
-  },
-  introButtonsCenter: {
-    paddingHorizontal: 24,
-    alignSelf: "stretch",
-    flexDirection: "row",
-    justifyContent: "center",
-  },
-  introButtonsWrap: {
-    flexGrow: 1,
-    maxWidth: 480,
   },
   paddedPage: {
     paddingTop: 64,
